Add FileDiffView to encapsulate file diff presentation
- Introduced a new FileDiffView.swift file which displays individual file diffs.
- Updated FileDiffsView to utilize the new FileDiffView for rendering each file diff instead of inline code.
- Updated Xcode project file to include new FileDiffView.swift.
